Add support for querying using GraphQL
Provide backend service that can take GraphQL queries and talk to
the database for data.

This allows support mobile app in which it uses GraphQL for data.

The `backend.ex` provides a set of callbacks that will be called
to interact with data. All database operations should be handled
here. The callbacks will be the REST methods: `create`, `retrieve`,
`update`, and `delete`.