Add mutation and viewer
- A sample mutation to create a trip
- A sample viewer object to carray all listing retrive queries and
the user if any.
- Add a plug to handle GraphQL calls from web, plus the GraphiQL to
test the queries.

Currently the mutation is not good. It's only there to demonstrate
the purpose. Ideally, that particular mutation needs to check for
permission, and relate the new trip to the user.

Also, the `Backend.create/2` callback is very naive. We need to
figure out a way for it to pattern match on different mutations.