Logout
Basic logout for now. Provides a mutation to delete a token, but it
does not do anything at the moment.

Ideally, when login, sign the token with a hash of the secret and the
user's `last_logout` timestamp. When verifying, load the user from the
`id` in the claims, and verify with the new hash. This should invalid
the old token because the "delete token" would update `last_logout`.